Moved all following team stuff to separate files
